requirements.txt
rasa==3.6.20
rasa-sdk
streamlit==1.33.0
dateparser==1.2.0
pandas==2.0.3
seaborn==0.13.2
matplotlib==3.5.3
requests==2.31.0
hdx-python-api
packaging==21.3
python-dateutil
text2digits==0.1.0
spacy
ruamel.yaml
unidecode==1.3.8
googletrans==4.0.0rc1
fasttext==0.9.2